requirements.txt
flask==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.22.3
Pillow==10.1.0
werkzeug==3.0.1
pytesseract==0.3.10
